Update dependencies, logging and alarms for 2021 season
Lots of updates to get everything working with Node 14 and improve
and logging and alarming that is now a couple of years old. Tests,
local ETL runs, deploys and scheduled ETL runs work. The only
remaining cranky element seems to be code coverage which doesn't
seem to work with .mjs files at present.